requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-dotenv>=1.0.0
mutagen>=1.47.0
Pillow>=10.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
pydantic[email]
alembic>=1.13.0
redis>=5.0.0
celery>=5.3.0
boto3>=1.34.0
stripe>=7.8.0